8931 Remove puppeteer flow, move getcartkey endpt to service
