+++ server/package.json
@@ -39,7 +39,6 @@
     "mocked-env": "^1.3.2",
     "moment": "^2.24.0",
     "nock": "^12.0.3",
-    "puppeteer": "^5.3.0",
     "reflect-metadata": "^0.1.13",
     "request": "^2.88.2",
     "rimraf": "^3.0.2",
